+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Reset failed login counter when accounts are unlocked (AUTHLIB-175)
+
 ## [4.2.1] - 2026-07-29
 
 ### Changed